requirements.txt
fastapi
uvicorn[standard]
python-multipart
numpy
scipy
librosa
soundfile
streamlit
requests
python-dotenv
streamlit-audiorecorder
